Plot chlorophyll and secchi data together with matrix outcomes

Plot chlorophyll and secchi data together with matrix outcomes

Usage

show_plotlypep(
  dat,
  bay_segment = c("1a", "1b", "2", "3"),
  yrrng = NULL,
  chllim = NULL,
  seclim = NULL
)

Arguments

dat

data.frame formatted from read_pepwq

bay_segment

chr string for the bay segment, one of "1a", "1b", "2", or "3"

yrrng

numeric for year range to plot

chllim

numeric vector of length two indicating range for the chlorophyll y-axis

seclim

numeric vector of length two indicating range for the secchi y-axis

Value

An interactive plotly object

Examples

show_plotlypep(rawdat)