Import dissolved oxygen data

Import dissolved oxygen data

Usage

read_pepdo(
  site = c("01304562", "01304200"),
  nms = NULL,
  startDate = "2013-01-01",
  endDate = "2021-12-31"
)

Arguments

site

chr string of site numbers for USGS stations

nms

optional chr string vector of names to reassign to site numbers

startDate

chr string of start date in YYYY-MM-DD

endDate

chr string of end date in YYYY-MM-DD

Value

data.frame

Details

Raw data are downloaded using the USGS dataRetrieval R package, this function is a simple wrapper to the readNWISuv function.

Note that downloading the station data with the default arguments may take a few minutes. Site are 01304562 for Peconic River, 01304200 for Orient Harbor.

Examples

if (FALSE) {
dodat <- read_pepdo(site = c('01304562', '01304200'), nms = c('Peconic R.', 'Orient Harbor'), 
     startDate = '2020-06-01', endDate = '2021-06-30')
dodat
}