Plot chlorophyll and secchi data together with matrix outcomes

Plot chlorophyll and secchi data together with matrix outcomes

Usage

show_plotlypep(
  dat,
  bay_segment = c("Western", "Central", "Eastern"),
  yrrng = NULL
)

Arguments

dat

data.frame formatted from read_pepwq

bay_segment

chr string for the bay segment, one of "Western", "Central", or "Eastern"

yrrng

numeric for year range to plot

Value

An interactive plotly object

Examples

show_plotlypep(rawdat)
#> Warning: `gather_()` was deprecated in tidyr 1.2.0.
#> Please use `gather()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.